requirements.txt
discord.py==2.5.0
youtube-dl==2021.12.17